+++ Foodsell/demo/src/main/java/com/example/demo/shops/ShopService.java
@@ -87,7 +87,7 @@ public class ShopService {
     }
     
     // Cập nhật rating cho shop
-    public Shop updateShopRating(int id, Double rating) {
+    public Shop updateShopRating(int id, Integer rating) {
         Shop shop = shopRepository.findById(id)
             .orElseThrow(() -> new RuntimeException("Shop not found with id: " + id));
         shop.setRating(rating);
